requirements.txt
dill
pandas
python-telegram-bot
numpy
catboost
scikit-learn